Remove `window.location.replace` make the back button work properly (#722)
* Remove `window.location.replace` make the back button work properly

While clicking around the marketing site of percy I noticed the
redirect to the docs page uses `window.location.replace`, which
doesn't add to the browser history. So when I used the back button I
ended up two pages ago rather than the page I was actually on before.

I did a grep of the project to see where this was used and it doesn't
_seem_ like the replace is a specific design decision, but I could be
wrong! If so we can split this out into two differnet methods to
handle replacing and setting.

* Add a new location method `replaceWindowLocation`

This splits the method `setWindowLocation` into two
methods. `setWindowLocation` stays and is now approartly named since
it sets the location and adds to the history.

A new method named `replaceWindowLocation` was added. That method does
what `setWindowLocation` did before. It replaced the window location
and does not add it to the history.

* Properly stub the right history util method in tests