Refactor project status icon (#838)
* Move project status icon to component

Project status is determined based on the most recent build for a project.
Sometimes we have that project in the store, but other times
we have to query for it. Previously this was done with a store.query
call in a computed project on a model. This was less than ideal, and
resulted in a backtracking render state that is no longer supported
by ember. This moves the crustiness and wraps it in a concurrency
task to clean it up.

* Remove crusty legacy code

* Refactor concurrency task

* Incorporate PR feedback